status pegawai change to client component

+++ src/app/master/status_pegawai/action.ts
@@ -2,70 +2,60 @@
 
 import { StatusPegawaiSchema } from "@_types/master/status_pegawai"
 import { setAuthorizeHeader } from "@helpers/index"
-import { API_URL } from "@lib/utils"
+import { API_URL, delay } from "@lib/utils"
 import axios from "axios"
-import { revalidateTag } from "next/cache"
 import { cookies } from "next/headers"
-import { redirect } from "next/navigation"
 
-/**
- * Saves a StatusPegawai object to the API.
- * @param _prevState - The previous state of the application.
- * @param formData - The FormData object containing the data to be saved.
- * @returns An object with an optional error property if the API request fails.
- * @throws An error if the API request is unsuccessful.
- */
-export const saveStatusPegawai = async (
-	_prevState: unknown,
-	formData: FormData,
-): Promise<{ error?: Record<string, string[]> }> => {
+export const saveStatusPegawai = async (formData: FormData) => {
 	const cookieList = cookies()
 	const headers = setAuthorizeHeader(cookieList)
 	try {
+		await delay(1000)
 		const validate = StatusPegawaiSchema.safeParse({
 			id: Number(formData.get("id")),
 			nama: formData.get("nama"),
 		})
-
 		if (!validate.success) {
-			return { error: validate.error.flatten().fieldErrors }
+			return {
+				success: false,
+				error: validate.error.flatten().fieldErrors,
+			}
 		}
 
 		const apiUrl =
 			validate.data.id > 0
 				? `${API_URL}/master/status-pegawai/${validate.data.id}`
 				: `${API_URL}/master/status-pegawai`
-
-		await axios.request({
+		const response = await axios.request({
 			method: validate.data.id > 0 ? "PUT" : "POST",
 			url: apiUrl,
 			data: formData,
 			headers: headers,
 		})
-		// biome-ignore lint/suspicious/noExplicitAny: <explanation>
+		return {
+			success: true,
+			data: response.data,
+		}
 	} catch (err: any) {
-		return { error: err.response.data }
+		return {
+			success: false,
+			error: { message: err.response?.data.message },
+		}
 	}
-
-	revalidateTag("status_pegawai")
-	redirect("/master/status_pegawai")
 }
 
-/**
- * Deletes a StatusPegawai object by its ID.
- *
- * @param _prevState - The previous state of the application.
- * @param formData - The FormData object containing the data to be deleted.
- * @returns A Promise that resolves to an object with a success property and an optional error property.
- */
-export const hapus = async (_prevState: unknown, formData: FormData) => {
-	const id = Number(formData.get("deleteRef")?.slice(7) || 0)
+export const hapus = async (formData: FormData) => {
+	const deleteRef = formData.get("deleteRef")?.toString() || ""
+	if (!deleteRef.startsWith("DELETE-"))
+		return { success: false, error: { message: "invalid data" } }
+	const id = Number(deleteRef.slice(7) || 0)
 	if (id <= 0) return { success: false, error: { message: "invalid data" } }
 
 	try {
 		await axios.delete(`${API_URL}/master/status-pegawai/${id}`, {
 			headers: setAuthorizeHeader(cookies()),
 		})
+		return { success: true }
 	} catch (err: any) {
 		return {
 			success: false,
@@ -74,7 +64,4 @@ export const hapus = async (_prevState: unknown, formData: FormData) => {
 			},
 		}
 	}
-
-	revalidateTag("status_pegawai")
-	redirect("/master/status_pegawai")
 }
